+++ feature/setting/src/main/java/com/boostcamp/dreamteam/dreamdiary/setting/component/SettingCategory.kt
@@ -1,0 +1,36 @@
+package com.boostcamp.dreamteam.dreamdiary.setting.component
+
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+
+@Composable
+internal fun SettingCategory(
+    text: String,
+    modifier: Modifier = Modifier,
+) {
+    Row(
+        modifier = modifier.padding(8.dp),
+        verticalAlignment = Alignment.CenterVertically,
+    ) {
+        Text(
+            text = text,
+            style = MaterialTheme.typography.bodyLarge,
+            color = MaterialTheme.colorScheme.primary,
+        )
+    }
+}
+
+@Composable
+@Preview
+private fun SettingCategoryPreview() {
+    SettingCategory(
+        text = "Settings",
+    )
+}
